added authorization

+++ functions/adminOtherFunctions/deleteWifi.function.js
@@ -7,6 +7,13 @@ module.exports.deleteWifi = functions.https.onCall(async (data, context) => {
   if (!context.auth) {
     return { message: "Authentication Required!", code: 401 };
   }
+  let userData = await db.collection("users").get();
+  if (userData.data().role != "admin") {
+    return {
+      message: "You are not authorized to perform this action",
+      code: 401
+    };
+  }
 
   const file = db.collection("wifiDetails").doc(data.uid);
   await file.delete();
